backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.50
langchain==0.3.0
langchain-openai==0.2.0
pydantic==2.9.0
pyyaml==6.0.2
